Scheduler: Allow callgraph tests to pass if rendering failed

+++ tests/test_scheduler.py
@@ -144,7 +144,8 @@ def test_scheduler_graph_simple(here, config, frontend):
     assert all((e[0].upper(), e[1].upper()) in vgraph.edges for e in expected_dependencies)
 
     cg_path.unlink()
-    cg_path.with_suffix('.pdf').unlink()
+    if cg_path.with_suffix('.pdf').exists():
+        cg_path.with_suffix('.pdf').unlink()
 
 
 @pytest.mark.skipif(importlib.util.find_spec('graphviz') is None, reason='Graphviz is not installed')
@@ -198,7 +199,8 @@ def test_scheduler_graph_partial(here, config, frontend):
     assert 'KERNELA' not in vgraph.nodes
 
     cg_path.unlink()
-    cg_path.with_suffix('.pdf').unlink()
+    if cg_path.with_suffix('.pdf').exists():
+        cg_path.with_suffix('.pdf').unlink()
 
 
 @pytest.mark.skipif(importlib.util.find_spec('graphviz') is None, reason='Graphviz is not installed')
@@ -237,7 +239,8 @@ def test_scheduler_graph_config_file(here, frontend):
     assert len(vgraph.edges) == 2
 
     cg_path.unlink()
-    cg_path.with_suffix('.pdf').unlink()
+    if cg_path.with_suffix('.pdf').exists():
+        cg_path.with_suffix('.pdf').unlink()
 
 
 @pytest.mark.skipif(importlib.util.find_spec('graphviz') is None, reason='Graphviz is not installed')
@@ -287,7 +290,8 @@ def test_scheduler_graph_blocked(here, config, frontend):
     assert len(vgraph.edges) == 4
 
     cg_path.unlink()
-    cg_path.with_suffix('.pdf').unlink()
+    if cg_path.with_suffix('.pdf').exists():
+        cg_path.with_suffix('.pdf').unlink()
 
 
 def test_scheduler_definitions(here, config, frontend):
@@ -396,7 +400,8 @@ def test_scheduler_graph_multiple_combined(here, config, frontend):
     assert all((e[0].upper(), e[1].upper()) in vgraph.edges for e in expected_dependencies)
 
     cg_path.unlink()
-    cg_path.with_suffix('.pdf').unlink()
+    if cg_path.with_suffix('.pdf').exists():
+        cg_path.with_suffix('.pdf').unlink()
 
 
 @pytest.mark.skipif(importlib.util.find_spec('graphviz') is None, reason='Graphviz is not installed')
@@ -452,7 +457,8 @@ def test_scheduler_graph_multiple_separate(here, config, frontend):
     assert all((e[0].upper(), e[1].upper()) in vgraph.edges for e in expected_dependenciesA)
 
     cg_path.unlink()
-    cg_path.with_suffix('.pdf').unlink()
+    if cg_path.with_suffix('.pdf').exists():
+        cg_path.with_suffix('.pdf').unlink()
 
     # Test second scheduler instance that holds the receiver items
     configB = config.copy()
@@ -489,7 +495,8 @@ def test_scheduler_graph_multiple_separate(here, config, frontend):
     assert ('EXT_DRIVER', 'EXT_KERNEL') in vgraphB.edges
 
     cg_path.unlink()
-    cg_path.with_suffix('.pdf').unlink()
+    if cg_path.with_suffix('.pdf').exists():
+        cg_path.with_suffix('.pdf').unlink()
 
 
 def test_scheduler_module_dependency(here, config, frontend):
